+++ server/rules/catch-all/create.test.ts
@@ -5,7 +5,7 @@ import {
   generateMockDocument,
   generateId,
   generateUserId,
-} from '../../test-helpers/contants';
+} from '../../test-helpers/constants';
 import {
   Firestore,
   setup,
