+++ server/rules/homestead/delete.test.ts
@@ -6,7 +6,7 @@ import {
   generateSecurityRecordOwner,
   generateUserId,
   membershipPath,
-} from '../../test-helpers/contants';
+} from '../../test-helpers/constants';
 import {
   Firestore,
   setup,
