+++ server/rules/homestead/read.test.ts
@@ -6,7 +6,7 @@ import {
   generateSecurityRecordAny,
   generateUserId,
   membershipPath,
-} from '../../test-helpers/contants';
+} from '../../test-helpers/constants';
 import {
   Firestore,
   setup,
