+++ server/rules/homestead/update.test.ts
@@ -10,7 +10,7 @@ import {
   generateSecurityRecordOwner,
   generateUserId,
   membershipPath,
-} from '../../test-helpers/contants';
+} from '../../test-helpers/constants';
 import {
   Firestore,
   setup,
